orchestrator_agent/requirements.txt
# Orchestrator Agent dependencies
google-adk[vertexai]>=1.0.0
google-genai>=1.0.0
python-dotenv>=1.0.0

